requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.4
colorama==0.4.6
coloredlogs==15.0.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imgbeddings==0.1.0
joblib==1.4.2
mpmath==1.3.0
numpy==2.1.2
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
